requirements.txt
Flask>=2.2,<3
Flask-SQLAlchemy>=3.0
passlib[bcrypt]>=1.7
cryptography>=41.0
boto3>=1.26
supabase>=2.0,<3
python-dotenv>=1.0
